hangar_stack/requirements.txt
# =============================================================================
# HangarStack - Aircraft Database with Kafka Integration
# Requirements File
# =============================================================================

# Core Web Framework
flask>=2.3.0

# Data Processing & Validation
jsonschema>=4.17.3
PyYAML>=6.0
pandas>=1.5.0

# HTTP & API
requests>=2.28.0
urllib3>=1.26.0
certifi>=2021.5.30

# Kafka Integration
kafka-python>=2.0.3
python-dotenv==1.0.0

# Real-time Features
redis==5.0.1
websockets==12.0

# Utilities
rich==13.7.0
six>=1.16.0

# Testing
pytest>=7.3.1

# =============================================================================
# Installation Instructions:
# pip install -r requirements.txt
# =============================================================================